+++ cogstat/cogstat_dialogs.py
@@ -664,6 +664,37 @@ class ylims_dialog(QtWidgets.QDialog, ylims.Ui_Dialog):
         return [self.lineEdit.text(), self.lineEdit_2.text()]
 
 
+from .ui import displayfactors_repeated
+class displayfactors_repeated_dialog(QtWidgets.QDialog, displayfactors_repeated.Ui_Dialog):
+    def __init__(self, parent=None):
+        QtWidgets.QDialog.__init__(self, parent)
+        self.setupUi(self)
+        self.setModal(True)
+        self.buttonBox.accepted.connect(self.accept)
+        self.buttonBox.rejected.connect(self.reject)
+        self.factor_x_listWidget.setDragDropMode(QtWidgets.QAbstractItemView.DragDropMode.DragDrop)
+        self.factor_x_listWidget.setDefaultDropAction(QtCore.Qt.MoveAction)
+        self.factor_x_listWidget.doubleClicked.connect(self.add_color)
+        self.factor_color_listWidget.setDragDropMode(QtWidgets.QAbstractItemView.DragDropMode.DragDrop)
+        self.factor_color_listWidget.setDefaultDropAction(QtCore.Qt.MoveAction)
+        self.factor_color_listWidget.doubleClicked.connect(self.remove_color)
+        self.add_color_button.clicked.connect(self.add_color)
+        self.remove_color_button.clicked.connect(self.remove_color)
+
+    def set_factors(self, factors=None):
+        self.factors = factors
+        _prepare_list_widgets(self.factor_x_listWidget, self.factors, [self.factor_color_listWidget])
+    def add_color(self):
+        _add_to_list_widget(self.factor_x_listWidget, self.factor_color_listWidget)
+    def remove_color(self):
+        _remove_item_from_list_widget(self.factor_x_listWidget, self.factor_color_listWidget, self.factors)
+    def read_parameters(self):
+        return ([str(self.factor_x_listWidget.item(i).text()) for i in range(self.factor_x_listWidget.count())] if
+                self.factor_x_listWidget.count() else [],
+                [str(self.factor_color_listWidget.item(i).text()) for i in range(self.factor_color_listWidget.count())] if
+                self.factor_color_listWidget.count() else [])
+
+
 from .ui import compare_vars
 class compare_vars_dialog(QtWidgets.QDialog, compare_vars.Ui_Dialog):
     def __init__(self, parent=None, names=[]):
@@ -682,10 +713,13 @@ class compare_vars_dialog(QtWidgets.QDialog, compare_vars.Ui_Dialog):
         self.removeVar.clicked.connect(self.remove_var)
         self.pushButton.clicked.connect(self.factorsButton_clicked)
         self.pushButton_2.clicked.connect(self.optionsButton_clicked)
+        self.pushButton_3.clicked.connect(self.displayfactorsButton_clicked)
 
         self.factors_dialog = factors_dialog(self)
         self.ylims_dialog = ylims_dialog(self)
+        self.displayfactors_repeated_dialog = displayfactors_repeated_dialog(self)
         self.factors = []
+        self.displayfactors = [[], []]
         self.ylims = [None, None]
 
         self.init_vars(names)
@@ -756,6 +790,11 @@ class compare_vars_dialog(QtWidgets.QDialog, compare_vars.Ui_Dialog):
                             _find_previous_item_position(self.source_listWidget, self.names, item.text()),
                             item.text())
 
+    def displayfactorsButton_clicked(self):
+        self.displayfactors_repeated_dialog.set_factors(factors=[factor[0] for factor in self.factors])
+        if self.displayfactors_repeated_dialog.exec_():
+            self.displayfactors = self.displayfactors_repeated_dialog.read_parameters()
+
     def optionsButton_clicked(self):
         if self.ylims_dialog.exec_():
             self.ylims = self.ylims_dialog.read_parameters()
@@ -765,10 +804,10 @@ class compare_vars_dialog(QtWidgets.QDialog, compare_vars.Ui_Dialog):
     def read_parameters(self):
         if len(self.factors) > 1:
             return [str(self.selected_listWidget.item(i).text().split(' :: ')[1]) for i in
-                    range(self.selected_listWidget.count())], self.factors, self.ylims
+                    range(self.selected_listWidget.count())], self.factors, self.displayfactors, self.ylims
         else:
             return [str(self.selected_listWidget.item(i).text()) for i in
-                    range(self.selected_listWidget.count())], self.factors, self.ylims
+                    range(self.selected_listWidget.count())], self.factors, self.displayfactors, self.ylims
 
 
 from .ui import compare_groups_single_case_slope
@@ -808,6 +847,47 @@ class compare_groups_single_case_slope_dialog(QtWidgets.QDialog, compare_groups_
                 str(self.spinBox.text()))
 
 
+from .ui import displayfactors_groups
+class displayfactors_groups_dialog(QtWidgets.QDialog, displayfactors_groups.Ui_Dialog):
+    def __init__(self, parent=None):
+        QtWidgets.QDialog.__init__(self, parent)
+        self.setupUi(self)
+        self.setModal(True)
+        self.buttonBox.accepted.connect(self.accept)
+        self.buttonBox.rejected.connect(self.reject)
+        self.factor_x_listWidget.setDragDropMode(QtWidgets.QAbstractItemView.DragDropMode.DragDrop)
+        self.factor_x_listWidget.setDefaultDropAction(QtCore.Qt.MoveAction)
+        self.factor_color_listWidget.setDragDropMode(QtWidgets.QAbstractItemView.DragDropMode.DragDrop)
+        self.factor_color_listWidget.setDefaultDropAction(QtCore.Qt.MoveAction)
+        self.factor_color_listWidget.doubleClicked.connect(self.remove_color)
+        self.factor_panel_listWidget.setDragDropMode(QtWidgets.QAbstractItemView.DragDropMode.DragDrop)
+        self.factor_panel_listWidget.setDefaultDropAction(QtCore.Qt.MoveAction)
+        self.factor_panel_listWidget.doubleClicked.connect(self.remove_panel)
+        self.add_color_button.clicked.connect(self.add_color)
+        self.remove_color_button.clicked.connect(self.remove_color)
+        self.add_panel_button.clicked.connect(self.add_panel)
+        self.remove_panel_button.clicked.connect(self.remove_panel)
+
+    def set_factors(self, factors=None):
+        self.factors = factors
+        _prepare_list_widgets(self.factor_x_listWidget, self.factors, [self.factor_color_listWidget, self.factor_panel_listWidget])
+    def add_color(self):
+        _add_to_list_widget(self.factor_x_listWidget, self.factor_color_listWidget)
+    def remove_color(self):
+        _remove_item_from_list_widget(self.factor_x_listWidget, self.factor_color_listWidget, self.factors)
+    def add_panel(self):
+        _add_to_list_widget(self.factor_x_listWidget, self.factor_panel_listWidget)
+    def remove_panel(self):
+        _remove_item_from_list_widget(self.factor_x_listWidget, self.factor_panel_listWidget, self.factors)
+    def read_parameters(self):
+        return ([str(self.factor_x_listWidget.item(i).text()) for i in range(self.factor_x_listWidget.count())] if
+                self.factor_x_listWidget.count() else [],
+                [str(self.factor_color_listWidget.item(i).text()) for i in range(self.factor_color_listWidget.count())] if
+                self.factor_color_listWidget.count() else [],
+                [str(self.factor_panel_listWidget.item(i).text()) for i in range(self.factor_panel_listWidget.count())] if
+                self.factor_panel_listWidget.count() else [])
+
+
 from .ui import compare_groups
 class compare_groups_dialog(QtWidgets.QDialog, compare_groups.Ui_Dialog):
     def __init__(self, parent=None, names=[]):
@@ -830,9 +910,12 @@ class compare_groups_dialog(QtWidgets.QDialog, compare_groups.Ui_Dialog):
         self.remove_group_button.clicked.connect(self.remove_group)
         self.pushButton.clicked.connect(self.on_slopeButton_clicked)
         self.pushButton_2.clicked.connect(self.optionsButton_clicked)
+        self.pushButton_3.clicked.connect(self.displayfactorsButton_clicked)
 
         self.slope_dialog = compare_groups_single_case_slope_dialog(self, names=names)
         self.ylims_dialog = ylims_dialog(self)
+        self.displayfactors_groups_dialog = displayfactors_groups_dialog(self)
+        self.displayfactors = [[], [], []]
         self.single_case_slope_SE, self.single_case_slope_trial_n = [], 0
         self.ylims = [None, None]
 
@@ -859,6 +942,12 @@ class compare_groups_dialog(QtWidgets.QDialog, compare_groups.Ui_Dialog):
         if self.slope_dialog.exec_():
             self.single_case_slope_SE, self.single_case_slope_trial_n = self.slope_dialog.read_parameters()
 
+    def displayfactorsButton_clicked(self):
+        self.displayfactors_groups_dialog.\
+            set_factors(factors=[str(self.group_listWidget.item(i).text()) for i in range(self.group_listWidget.count())])
+        if self.displayfactors_groups_dialog.exec_():
+            self.displayfactors = self.displayfactors_groups_dialog.read_parameters()
+
     def optionsButton_clicked(self):
         if self.ylims_dialog.exec_():
             self.ylims = self.ylims_dialog.read_parameters()
@@ -868,6 +957,7 @@ class compare_groups_dialog(QtWidgets.QDialog, compare_groups.Ui_Dialog):
     def read_parameters(self):
         return ([str(self.selected_listWidget.item(i).text()) for i in range(self.selected_listWidget.count())],
                 [str(self.group_listWidget.item(i).text()) for i in range(self.group_listWidget.count())],
+                self.displayfactors,
                 self.single_case_slope_SE, int(self.single_case_slope_trial_n), self.ylims)